@@ -1,15 +1,17 @@
 [build-system]
-requires = ["maturin>=0.11.5"]
-build-backend = "maturin"
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
 
-[project]
+[tool.poetry]
 name = "flight-fusion"
-description = "Client to filight fusion service"
+version = "0.2.0"
+description = "Client for interacting with flight fusion service"
 readme = "README.md"
-license = { file = "LICENSE.txt" }
-requires-python = ">=3.6"
+authors = ["Robert Pack <robstar.pack@gmail.com>"]
+license = "MIT"
+repository = "https://github.com/roeap/flight-fusion"
 keywords = ["deltalake", "delta", "arrow", "datafusion", "arrow-flight"]
-classifier = [
+classifiers = [
   "Development Status :: 2 - Pre-Alpha",
   "Intended Audience :: Developers",
   "Intended Audience :: Science/Research",
@@ -27,32 +29,31 @@ classifier = [
   "Programming Language :: Python",
   "Programming Language :: Rust",
 ]
-dependencies = [
-  "betterproto==2.0.0b4",
-  "loguru>=0.5",
-  "pandas>=1",
-  "pyarrow>=4",
-  "pydantic[dotenv]>=1",
-  "SQLAlchemy>=1.4",
-  "toml>=0.10",
-  "typer>=0.4",
-]
 
-[project.optional-dependencies]
-devel = [
-  "black",
-  "isort",
-  "pytest",
-  "pytest-mock",
-  "pytest-cov",
-  "pytest-datadir>=1.3.1",
-]
-docker = ["python-on-whales>=0.36"]
+[tool.poetry.dependencies]
+python = "^3.8"
+betterproto = "2.0.0b4"
+loguru = ">=0.5"
+pandas = ">=1"
+pyarrow = ">=4"
+pydantic = { version = ">=1", extras = ["dotenv"] }
+SQLAlchemy = ">=1.4"
+toml = ">=0.10"
+typer = ">=0.4"
 
-[project.urls]
-repository = "https://github.com/roeap/flight-fusion"
+python-on-whales = { version = ">=0.36", optional = true }
 
-[project.scripts]
+[tool.poetry.dev-dependencies]
+pytest = "^6.2.5"
+pytest-azurepipelines = "^1"
+pytest-datadir = "^1.3.1"
+pytest-cov = "^3"
+pytest-mock = "*"
+
+[tool.poetry.extras]
+docker = ["python-on-whales"]
+
+[tool.poetry.scripts]
 fusion = "flight_fusion.cli:app"
 
 [tool.isort]
