@@ -1,32 +1,29 @@
 import numpy as np
 import pandas as pd
-import pytest
 
 from flight_fusion import FusionServiceClient
 from flight_fusion.ipc.v1alpha1 import SaveMode
 
 
-@pytest.mark.integration
 def test_roundtrip(ff_client: FusionServiceClient):
     np.random.seed(42)
     df = pd.DataFrame(np.random.randn(5, 3), columns=["col1", "col2", "col3"])
 
-    fds = ff_client.get_dataset_client(name="new_dataset", areas=["asd", "fgh"])
-    fds.write_into(df)
+    fds = ff_client.get_dataset_client(name="table", areas=["test_roundtrip"])
+    fds.write_into(df, SaveMode.SAVE_MODE_OVERWRITE)
 
     df_loaded = fds.load().to_pandas()
 
     assert df.equals(df_loaded)
 
 
-@pytest.mark.integration
 def test_save_mode(ff_client: FusionServiceClient):
     np.random.seed(42)
     df = pd.DataFrame(np.random.randn(5, 3), columns=["col1", "col2", "col3"])
 
-    fds = ff_client.get_dataset_client(name="new_dataset", areas=["asd", "fgh"])
+    fds = ff_client.get_dataset_client(name="table", areas=["test_save_mode"])
 
-    fds.write_into(df)
+    fds.write_into(df, SaveMode.SAVE_MODE_OVERWRITE)
     df_loaded = fds.load().to_pandas()
     assert df_loaded.shape == (5, 3)
 
