@@ -1,27 +1,31 @@
-use super::{
-    error::*, stats, utils::*, writer::*, AreaStore, DATA_FOLDER_NAME, DEFAULT_READ_BATCH_SIZE,
-};
+use super::{error::*, stats, utils::*, writer::*, AreaStore, DATA_FOLDER_NAME};
 use arrow_deps::arrow::record_batch::*;
 use arrow_deps::datafusion::parquet::{
-    arrow::{ArrowReader, ParquetFileArrowReader},
+    arrow::ParquetFileArrowReader,
     file::serialized_reader::{SerializedFileReader, SliceableCursor},
 };
 use async_trait::async_trait;
 use flight_fusion_ipc::{
     area_source_reference::Table as TableReference, AreaSourceReference, SaveMode,
 };
+use object_store::path::Path;
 use object_store::{path::ObjectStorePath, ObjectStoreApi};
 use std::path::PathBuf;
 use std::sync::Arc;
 
 pub struct DefaultAreaStore {
     object_store: Arc<object_store::ObjectStore>,
+    root_path: String,
 }
 
 impl DefaultAreaStore {
     pub fn new(root: impl Into<PathBuf>) -> Self {
-        let object_store = Arc::new(object_store::ObjectStore::new_file(root));
-        Self { object_store }
+        let buf: PathBuf = root.into();
+        let object_store = Arc::new(object_store::ObjectStore::new_file(buf.clone()));
+        Self {
+            object_store,
+            root_path: format!("{}", buf.to_str().unwrap()),
+        }
     }
 
     pub fn new_azure(
@@ -29,13 +33,22 @@ impl DefaultAreaStore {
         access_key: impl Into<String>,
         container_name: impl Into<String>,
     ) -> Result<Self> {
+        let container: String = container_name.into();
         let object_store = Arc::new(object_store::ObjectStore::new_microsoft_azure(
             account,
             access_key,
-            container_name,
+            container.clone(),
             false,
         )?);
-        Ok(Self { object_store })
+        Ok(Self {
+            object_store,
+            root_path: format!("adls2://{}", container),
+        })
+    }
+
+    pub fn get_full_table_path(&self, source: &AreaSourceReference) -> Result<String> {
+        let location = self.get_table_location(source)?;
+        Ok(format!("{}/{}", self.root_path, location.to_raw()))
     }
 }
 
@@ -43,6 +56,13 @@ impl DefaultAreaStore {
 impl AreaStore for DefaultAreaStore {
     fn object_store(&self) -> Arc<object_store::ObjectStore> {
         self.object_store.clone()
+    }
+
+    fn get_path_from_raw(&self, raw: String) -> Path {
+        let trimmed_raw = raw
+            .trim_start_matches(&self.root_path)
+            .trim_start_matches('/');
+        self.object_store.path_from_raw(trimmed_raw)
     }
 
     fn get_table_location(&self, source: &AreaSourceReference) -> Result<object_store::path::Path> {
@@ -98,12 +118,7 @@ impl AreaStore for DefaultAreaStore {
         let files = self.get_location_files(location).await?;
         let mut batches = Vec::new();
         for file in files {
-            let mut reader = self.get_arrow_reader(&file).await?;
-            let batch_reader = reader.get_record_reader(DEFAULT_READ_BATCH_SIZE).unwrap();
-            let mut file_batches = batch_reader
-                .into_iter()
-                .collect::<std::result::Result<Vec<_>, _>>()?;
-            batches.append(&mut file_batches);
+            batches.append(&mut self.read_file(&file).await?);
         }
         Ok(batches)
     }
