@@ -1,17 +1,13 @@
-use super::*;
+use super::{utils::create_response_stream, *};
 use crate::error::{FusionServiceError, Result};
 use area_store::store::AreaStore;
-use arrow_deps::{
-    arrow::{ipc::writer::IpcWriteOptions, record_batch::RecordBatch},
-    datafusion::prelude::{ExecutionConfig, ExecutionContext},
-};
-use arrow_flight::{FlightData, SchemaAsIpc};
+use arrow_deps::datafusion::prelude::{SessionConfig, SessionContext};
+use arrow_flight::FlightData;
 use async_trait::async_trait;
 use flight_fusion_ipc::{
     command_execute_query::Context as QueryContext, CommandExecuteQuery, CommandReadDataset,
     CommandSqlOperation,
 };
-use tonic::Status;
 
 #[async_trait]
 impl DoGetHandler<CommandSqlOperation> for FusionActionHandler {
@@ -19,8 +15,8 @@ impl DoGetHandler<CommandSqlOperation> for FusionActionHandler {
         &self,
         ticket: CommandSqlOperation,
     ) -> Result<BoxedFlightStream<FlightData>> {
-        let config = ExecutionConfig::new().with_information_schema(true);
-        let mut ctx = ExecutionContext::with_config(config);
+        let config = SessionConfig::new().with_information_schema(true);
+        let ctx = SessionContext::with_config(config);
         ctx.register_catalog("catalog", self.catalog.clone());
 
         // execute the query
@@ -54,7 +50,7 @@ impl DoGetHandler<CommandExecuteQuery> for FusionActionHandler {
         &self,
         ticket: CommandExecuteQuery,
     ) -> Result<BoxedFlightStream<FlightData>> {
-        let mut ctx = ExecutionContext::new();
+        let mut ctx = SessionContext::new();
         match ticket.context {
             Some(QueryContext::Source(source)) => {
                 self.register_source(&mut ctx, &source).await?;
@@ -68,38 +64,4 @@ impl DoGetHandler<CommandExecuteQuery> for FusionActionHandler {
         };
         create_response_stream(ctx.sql(&ticket.query).await?.collect().await?).await
     }
-}
-
-async fn create_response_stream(
-    results: Vec<RecordBatch>,
-) -> Result<BoxedFlightStream<FlightData>> {
-    if results.is_empty() {
-        return Err(FusionServiceError::NoReturnData(
-            "There were no results from ticket".to_string(),
-        ));
-    }
-    let schema = results[0].schema();
-    // TODO get rid of all the panics
-    let options = IpcWriteOptions::default();
-    let schema_flight_data = SchemaAsIpc::new(&schema, &options).into();
-
-    let mut flights: Vec<std::result::Result<FlightData, Status>> = vec![Ok(schema_flight_data)];
-    let mut batches: Vec<std::result::Result<FlightData, Status>> = results
-        .iter()
-        .flat_map(|batch| {
-            let (flight_dictionaries, flight_batch) =
-                arrow_flight::utils::flight_data_from_arrow_batch(batch, &options);
-            flight_dictionaries
-                .into_iter()
-                .chain(std::iter::once(flight_batch))
-                .map(Ok)
-        })
-        .collect();
-
-    // append batch vector to schema vector, so that the first message sent is the schema
-    flights.append(&mut batches);
-
-    let output = futures::stream::iter(flights);
-
-    Ok(Box::pin(output) as BoxedFlightStream<FlightData>)
 }
